added with-builtinmodules
Can you please try it with this changed configuration options?

I removed "--disable-silent-rules", because it was only handy for development purposes, but serves no real purpose apart from that. 

I also set syzextra and gfanlib to builtinmodules, which gets rid of the need to load a shared module from inside Singular to use them.